refactor: pull clipboard and subprocess interactions out to a trait
