+++ src/dot/text_object.rs
@@ -11,7 +11,7 @@ use crate::{
 /// A vim-like text object which can be used to manipulate the current Dot in a Buffer
 #[allow(dead_code)]
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub(crate) enum TextObject {
+pub enum TextObject {
     Arr(Arrow),
     BufferEnd,
     BufferStart,
