+++ src/input.rs
@@ -13,7 +13,7 @@ use std::{
 
 /// An input event that can be processed by the editor event loop
 #[derive(Debug)]
-pub(crate) enum Event {
+pub enum Event {
     /// A [Message] received from the virtual filesystem interface
     Message(Message),
     /// An [Input] from the user
